+++ src/layer.cpp
@@ -293,8 +293,8 @@ int Layer::ExportMIDI(ArrayPtrVoid *params)
     MidiFile *midiFile = static_cast<MidiFile *>((*params).at(0));
     int *midiTrack = static_cast<int *>((*params).at(1));
     MeterSig **currentMeterSig = static_cast<MeterSig **>((*params).at(2));
-    int *currentMeasureTime = static_cast<int *>((*params).at(3));
-    int *totalTime = static_cast<int *>((*params).at(4));
+    double *currentMeasureTime = static_cast<double *>((*params).at(3));
+    double *totalTime = static_cast<double *>((*params).at(4));
 
     // Here we set the current meterSig
     (*currentMeterSig) = GetCurrentMeterSig();
